+++ src/main/java/com/cggov/labour/ruralservice/repository/ApplicantInformationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.cggov.labour.ruralservice.datamodel.ApplicantInformationData;
 
 
-public interface ApplicantInformationRepository extends JpaRepository<ApplicantInformationData, Long> {
+public interface ApplicantInformationRepository extends JpaRepository<ApplicantInformationData, Integer> {
 
 }
